Python/requirements.txt
ultralytics
flask
waitress
pillow
roboflow

fastapi
uvicorn
pydantic
langchain
langchain-google-genai
langchain-experimental
langchain-groq
langchain-cohere
llama-index
llama-index-llms-openrouter
pandas
yt-dlp
beautifulsoup4
requests
faiss-cpu
cohere
google-generativeai
openai
llama-index
llama-index-experimental
langchain_together
google
reportlab